requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
aiofiles==23.2.1
    # via
    #   -r requirements.in
    #   aiogram
aiogram==3.2.0
    # via -r requirements.in
aiohttp==3.9.1
    # via
    #   -r requirements.in
    #   aiogram
    #   black
aiosignal==1.3.1
    # via
    #   -r requirements.in
    #   aiohttp
annotated-types==0.6.0
    # via
    #   -r requirements.in
    #   pydantic
anyio==4.1.0
    # via
    #   -r requirements.in
    #   httpx
asgiref==3.7.2
    # via
    #   -r requirements.in
    #   django
attrs==23.1.0
    # via
    #   -r requirements.in
    #   aiohttp
black==23.12.0
    # via -r requirements.in
build==1.0.3
    # via
    #   -r requirements.in
    #   pip-tools
certifi==2023.11.17
    # via
    #   -r requirements.in
    #   aiogram
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via
    #   -r requirements.in
    #   cryptography
charset-normalizer==3.3.2
    # via
    #   -r requirements.in
    #   requests
click==8.1.7
    # via
    #   -r requirements.in
    #   black
    #   pip-tools
colorama==0.4.6
    # via -r requirements.in
cryptography==41.0.7
    # via
    #   -r requirements.in
    #   social-auth-core
dawg-python==0.7.2
    # via
    #   -r requirements.in
    #   pymorphy2
defusedxml==0.8.0rc2
    # via
    #   -r requirements.in
    #   python3-openid
    #   social-auth-core
django==5.0
    # via
    #   -r requirements.in
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   djoser
    #   social-auth-app-django
django-filter==23.5
    # via -r requirements.in
django-profanity-filter==0.2.1
    # via -r requirements.in
django-templated-mail==1.1.1
    # via
    #   -r requirements.in
    #   djoser
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.3.1
    # via
    #   -r requirements.in
    #   djoser
djoser==2.2.2
    # via -r requirements.in
dnspython==2.4.2
    # via
    #   -r requirements.in
    #   email-validator
docopt==0.6.2
    # via
    #   -r requirements.in
    #   pymorphy2
email-validator==2.1.0.post1
    # via -r requirements.in
et-xmlfile==1.1.0
    # via
    #   -r requirements.in
    #   openpyxl
frozenlist==1.4.0
    # via
    #   -r requirements.in
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via
    #   -r requirements.in
    #   httpcore
httpcore==1.0.2
    # via
    #   -r requirements.in
    #   httpx
httpx==0.25.2
    # via -r requirements.in
idna==3.6
    # via
    #   -r requirements.in
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   yarl
inflection==0.5.1
    # via
    #   -r requirements.in
    #   django-profanity-filter
iniconfig==2.0.0
    # via
    #   -r requirements.in
    #   pytest
install==1.3.5
    # via -r requirements.in
isort==5.13.1
    # via -r requirements.in
magic-filter==1.0.12
    # via
    #   -r requirements.in
    #   aiogram
multidict==6.0.4
    # via
    #   -r requirements.in
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via
    #   -r requirements.in
    #   black
oauthlib==3.2.2
    # via
    #   -r requirements.in
    #   requests-oauthlib
    #   social-auth-core
openpyxl==3.1.2
    # via -r requirements.in
packaging==23.2
    # via
    #   -r requirements.in
    #   black
    #   build
    #   pytest
pathspec==0.12.1
    # via
    #   -r requirements.in
    #   black
pip-tools==7.3.0
    # via -r requirements.in
platformdirs==4.1.0
    # via
    #   -r requirements.in
    #   black
pluggy==1.3.0
    # via
    #   -r requirements.in
    #   pytest
pyaes==1.6.1
    # via
    #   -r requirements.in
    #   pyrogram
    #   telethon
pyasn1==0.5.1
    # via
    #   -r requirements.in
    #   rsa
pycparser==2.21
    # via
    #   -r requirements.in
    #   cffi
pydantic==2.5.2
    # via
    #   -r requirements.in
    #   aiogram
pydantic-core==2.14.5
    # via
    #   -r requirements.in
    #   pydantic
pyjwt==2.8.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   social-auth-core
pymorphy2==0.8
    # via -r requirements.in
pymorphy2-dicts==2.4.393442.3710985
    # via
    #   -r requirements.in
    #   pymorphy2
pyproject-hooks==1.0.0
    # via
    #   -r requirements.in
    #   build
pyrogram==2.0.106
    # via -r requirements.in
pysocks==1.7.1
    # via
    #   -r requirements.in
    #   pyrogram
pytest==7.4.3
    # via
    #   -r requirements.in
    #   pytest-asyncio
pytest-asyncio==0.23.2
    # via -r requirements.in
python-dotenv==1.0.0
    # via -r requirements.in
python3-openid==3.2.0
    # via
    #   -r requirements.in
    #   social-auth-core
pytz==2023.3.post1
    # via
    #   -r requirements.in
    #   djangorestframework
requests==2.31.0
    # via
    #   -r requirements.in
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==1.3.1
    # via
    #   -r requirements.in
    #   social-auth-core
rsa==4.9
    # via
    #   -r requirements.in
    #   telethon
sniffio==1.3.0
    # via
    #   -r requirements.in
    #   anyio
    #   httpx
social-auth-app-django==5.4.0
    # via
    #   -r requirements.in
    #   djoser
social-auth-core==4.5.1
    # via
    #   -r requirements.in
    #   social-auth-app-django
sqlparse==0.4.4
    # via
    #   -r requirements.in
    #   django
telethon==1.33.1
    # via -r requirements.in
typing-extensions==4.9.0
    # via
    #   -r requirements.in
    #   aiogram
    #   pydantic
    #   pydantic-core
tzdata==2023.3
    # via -r requirements.in
urllib3==2.1.0
    # via
    #   -r requirements.in
    #   requests
wheel==0.42.0
    # via pip-tools
yarl==1.9.4
    # via
    #   -r requirements.in
    #   aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
